xMsg: don't wait too much after subscription
No need to wait 100 ms just after subscribing, because the SUB socket is
being checked with the control socket to see if the subscription is
registered.

But that only means that the subscription was propagated to
the proxy control thread. The actual requested PUB socket may still be
missing the subscription, so a small delay is still needed.

Unfortunately, this new delay could be too short anyway, and the
response would not be sent. A final solution to this problem is not
reached yet.