xMsg: fix generated replyTo topic on syncPublish
A random number between 0 and 100 is creating many conflicts when a
unique replyTo topic must be generated, and those response messages are
being lost.

Replace the random number generation with an identifier plus a unique
sequence of 6 digits:

    ret:<id>:<prefix_digit><6_digits>

The identifier is unique for the xMsg agent, based on a hash of its
address and name. Thus, it will be the same for every request sent by
the agent. This ensure different actors will have a different
identifier.

The sequence of digits is global in the C++ process. To ensure every
single requests uses a different topic, every time a replyTo is
required, the next number in the sequence will be used. The sequence
goes from 2000000 to 2999999 and then repeats. 6 digits seems good
enough, because 1 million of concurrent syncPublish messages is not even
considered.