perf: fix throughput test not sending all messages
The context must be destroyed in order to ensure publication of all
messages still on queue.

But on OSX (where we can not use a proper thread_local implementation),
the destructor of the thread pool for the main thread is never called,
due to the main thread not calling pthread_exit [1].

Since the thread pool uses a shared pointer to the global context, the
context is never destroyed, and enqueued messages are not published.

Use a global variable trick to force calling the destructor of the
thread pool, which in turns allows the context to be destroyed.

[1]: http://stackoverflow.com/a/36438704