context: encapsulate 0MQ context
Do not use zmq::context_t in any public class.

Provide an internal wrapper xmsg::detail::Context that encapsulates the
zmq::context_t object with some utility methods, and use the Pimpl idiom
in the xmsg::Context class with a pointer to the wrapper.

The idea is to remove the public dependency on the zmq.hpp header, and
keep it internal (it may even be discarded with custom, more efficient
classes). Commit 8bd4661e3024 ("connection: refactor customization of
sockets") already removed the zmq::socket_t from the xMsg public API.

Some clients may need to use extra contexts and not just the global
singleton, so a factory method to create new contexts has been added.
These new contexts should be able to be moved if required for exclusive
access (for example, the xmsg::Proxy must use its own context, because
the context must be closed in order to stop the proxy).

But using and moving xmsg::Contexts objects introduces the possibility
of moving the global singleton, and other threads using the singleton
will fail.

To prevent this, wrap the contexts returned by the static methods into
an appropriated smart pointer: a std::shared_pointer for the singleton
and a std::unique_ptr for new contexts. Since the xmsg::Context class is
always contained in a smart pointer, its copy and move constructors have
been disabled. Now the singleton itself cannot be moved and left in an
undefined state, but extra contexts can be created, moved and destroyed
if necessary.