zhelper: rename 'core' sub-namespace to 'detail'
It is the convention for the name of the sub-namespace with
implementation details, and when those classes are visible to the user.

Move from 'core' in zhelper.h because it will contain an implementation
context class to encapsulate the zmq.hpp context, and this class will
need to be declared in a public header.