connection: use same setup for all connections in a pool
Do not support a setup parameter when requesting a connection with
get_connection, because it doesn't ensure that a connection will be
created. If it is not a new connection but a reused one, then it will
have a different setup, which is confusing.

To avoid this situation, make all connections in a single pool share the
same setup.