Update build instructions to latest protobuf version on macOS
Commit ff2d3edfcba0 ("Update build instructions to new Homebrew protobuf
formula") introduced a keg-only formula to compile with protobuf 2.6,
but using a keg-only formula is just too convoluted, and CMake can still
fail if the formula is not linked.

Prefer installing the latest version of protobuf (version 3). It seems
to work fine with the Java version compiled with 2.6, because the
messages are still compiled from proto2 syntax.